% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot}
\alias{forgot}
\title{Retrieve tibble with function documentation for a specified package}
\usage{
forgot(pkg, keyword = NULL, selected = NULL, formatted = T, interactive = F)
}
\arguments{
\item{pkg}{string with installed R package name}

\item{keyword}{optional string to filter the dataframe on for any mention of}

\item{selected}{optional string vector of columns to search on, null by default}

\item{formatted}{boolean should it coerce columns to char, true by default}

\item{interactive}{optional boolean to get a reactable HTML table}
}
\value{
tibble with function documentation info
}
\description{
Retrieve tibble with function documentation for a specified package
}
\examples{
forgot::forgot("stringr")
forgot::forgot("dplyr", "count")
forgot::forgot(dplyr, "count")
}
