% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot2}
\alias{forgot2}
\title{Get forgot tibble with only first two fields populated by default}
\usage{
forgot2(pkg, keyword = NULL, selected = c())
}
\arguments{
\item{pkg}{string with installed R package name}

\item{keyword}{optional string to filter the dataframe on for any mention of}

\item{selected}{optional string vector of columns to add on to search}
}
\value{
tibble with function documentation info
}
\description{
Get forgot tibble with only first two fields populated by default
}
\examples{
forgot2("stringr")
forgot2(stringr)
forgot2("dplyr", "count")
forgot2(dplyr, "count")
}
