% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot_exmpls}
\alias{forgot_exmpls}
\title{Print examples for an R function of interest in console}
\usage{
forgot_exmpls(pkg, function_name)
}
\arguments{
\item{pkg}{string with installed R package name}

\item{function_name}{function name in R package}
}
\description{
Print examples for an R function of interest in console
}
\examples{
forgot_exmpls("dplyr", "count")
}
