% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot_fx}
\alias{forgot_fx}
\title{Get tibble with roxygen2 fields for a specific function of interest}
\usage{
forgot_fx(pkg, function_name, field = NULL, print = TRUE)
}
\arguments{
\item{pkg}{string installed R package name}

\item{function_name}{string function name in R package}

\item{field}{string roxygen2 field of interest, NULL by default}

\item{print}{logical on if you should print data frame, TRUE by default}
}
\value{
forgot tibble with roxygen2 field info for pkg or function
}
\description{
Get tibble with roxygen2 fields for a specific function of interest
}
\examples{
forgot_fx("dplyr", "count", "usage", print = FALSE)
forgot_fx("dplyr", "count", "usage", print = TRUE)
}
