% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot_params}
\alias{forgot_params}
\title{Print parameters for an R function of interest in console}
\usage{
forgot_params(pkg, function_name)
}
\arguments{
\item{pkg}{string with installed R package name}

\item{function_name}{function name in R package}
}
\description{
Print parameters for an R function of interest in console
}
\examples{
forgot_params("dplyr", "count")
}
