% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot_usg}
\alias{forgot_usg}
\title{Print usage for an R function of interest in console}
\usage{
forgot_usg(pkg, function_name, write = FALSE)
}
\arguments{
\item{pkg}{string with installed R package name}

\item{function_name}{function name in R package}

\item{write}{logical write to RStudio rmd file, FALSE by default}
}
\description{
Print usage for an R function of interest in console
}
\examples{
forgot_usg("dplyr", "count")
forgot_usg(dplyr, count)
\dontrun{
forgot_usg(dplyr, count, write = TRUE)
}
}
