% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgotAddin.R
\name{forgotAddin}
\alias{forgotAddin}
\title{Forgot Rstudio Addin to Search Package Documentation}
\usage{
forgotAddin(selected_pkg = NULL)
}
\arguments{
\item{selected_pkg}{string optional package name to search on}
}
\description{
Creates the miniUI for the forgot RStudio Addin
}
\examples{
\dontrun{
forgotAddin("dplyr")
}
}
