% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{new_file_write}
\alias{new_file_write}
\title{Create new RStudio Rmarkdown document with content of interest}
\usage{
new_file_write(content, type = "rmarkdown")
}
\arguments{
\item{content}{vector with text to include in file}

\item{type}{type of file to create in RStudio IDE, RMarkdown by default}
}
\description{
Create new RStudio Rmarkdown document with content of interest
}
