% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_html_table}
\alias{simple_html_table}
\title{Get simple reactable HTML table}
\usage{
simple_html_table(data)
}
\arguments{
\item{data}{tibble dataset for reactable}
}
\value{
reactable HTML object
}
\description{
Get simple reactable HTML table
}
\examples{
simple_html_table(iris)
}
