% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_reactable_table}
\alias{simple_reactable_table}
\title{Get simple reactable table}
\usage{
simple_reactable_table(data, shiny_vers = F)
}
\arguments{
\item{data}{tibble with dataset}

\item{shiny_vers}{option to return shiny version used in `simple_html_table()`}
}
\value{
reactable object
}
\description{
Get simple reactable table
}
\examples{
simple_reactable_table(iris)
}
