% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgot.R
\name{forgot}
\alias{forgot}
\title{Retrieve tibble with function documentation for a specified package}
\usage{
forgot(pkg, formatted = TRUE)
}
\arguments{
\item{pkg}{string with installed R package name}

\item{formatted}{boolean should it coerce columns to char, true by default}
}
\value{
tibble with function documentation info
}
\description{
Retrieve tibble with function documentation for a specified package
}
\examples{
forgot::forgot("stringr")
}
